Implement JDBCBackend.ts_delete
